Use timeout, if available, and fallback to ping
Given that timeout.exe is available on most systems, it is preferable over ping.exe, but the former is not available on Windows XP. Accordingly, the most robust approach might be to use timeout, if available, and fallback to ping. The stream redirection will squelch the error that would be emitted otherwise when timeout.exe is unavailable (and it should squelch the output from ping.exe, too).